+++ ckanext/toolbelt/cli/__init__.py
@@ -2,6 +2,7 @@ import click
 
 from .make import make
 from .dev import dev
+from .search_index import search_index
 
 @click.group(short_help="Variety of useful commands.")
 def toolbelt():
@@ -10,3 +11,4 @@ def toolbelt():
 
 toolbelt.add_command(make)
 toolbelt.add_command(dev)
+toolbelt.add_command(search_index)